+++ R/utils.R
@@ -1,7 +1,7 @@
 # PMML: Predictive Model Markup Language
 #
 # Copyright (c) 2009-2016, Zementis, Inc.
-# Copyright (c) 2016-2019, Software AG, Darmstadt, Germany and/or Software AG
+# Copyright (c) 2016-2020, Software AG, Darmstadt, Germany and/or Software AG
 # USA Inc., Reston, VA, USA, and/or its subsidiaries and/or its affiliates
 # and/or their licensors.
 #
@@ -32,8 +32,9 @@
 
 
 
-.markupSpecials <- function(x)
+.markupSpecials <- function(x) {
   gsub("<", "&lt;", gsub(">", "&gt;", gsub("&", "&amp;", x)))
+}
 
 
 .generateCopyright <- function() {