+++ R/xform_norm_discrete.R
@@ -1,7 +1,7 @@
 # PMML: Predictive Model Markup Language
 #
 # Copyright (c) 2009-2016, Zementis, Inc.
-# Copyright (c) 2016-2019, Software AG, Darmstadt, Germany and/or Software AG
+# Copyright (c) 2016-2020, Software AG, Darmstadt, Germany and/or Software AG
 # USA Inc., Reston, VA, USA, and/or its subsidiaries and/or its affiliates
 # and/or their licensors.
 #
@@ -26,7 +26,6 @@
 #' the input variable to be transformed.
 #' @param input_var The input variable name in the data on which the
 #' transformation is to be applied.
-#' @param inputVar Deprecated.
 #' @param map_missing_to Value to be given to the transformed variable if the
 #' value of the input variable is missing.
 #' @param \dots Further arguments passed to or from other methods.
@@ -103,19 +102,8 @@
 xform_norm_discrete <- function(wrap_object,
                                 xform_info = NA,
                                 input_var = NA,
-                                inputVar,
                                 map_missing_to = NA,
                                 ...) {
-
-  # Deprecated argument.
-  if (!missing(inputVar)) {
-    warning("argument inputVar is deprecated; please use input_var instead.",
-      call. = FALSE
-    )
-    input_var <- inputVar
-  }
-
-
   map <- NULL
   colmn <- NULL
   newrow <- NULL