+++ tests/testthat/test_schema_validation.R
@@ -137,8 +137,8 @@ zmz_transform_audit <- function(box_obj) {
 }
 
 
-schema <- XML::xmlSchemaParse("pmml-4-4_xslt_20190830_10.5.0.0.xsd")
-
+# schema <- XML::xmlSchemaParse("pmml-4-4_xslt_20190830_10.5.0.0.xsd")
+schema <- XML::xmlSchemaParse("pmml-4-4.xsd")
 
 test_that("TimeSeries/Arima PMML validates against schema", {
   fit <- Arima(WWWusage, order = c(1, 0, 1))
@@ -155,6 +155,18 @@ test_that("TimeSeries/Arima PMML validates against schema", {
 
   fit <- Arima(AirPassengers, order = c(0, 1, 1), seasonal = c(0, 1, 1))
   expect_equal(validate_pmml(pmml(fit), schema), 0)
+
+  fit <- Arima(JohnsonJohnson, order = c(0, 1, 0), seasonal = c(0, 1, 2))
+  expect_equal(validate_pmml(pmml(fit, exact_least_squares = TRUE), schema), 0)
+
+  fit <- Arima(AirPassengers, order = c(0, 1, 1), seasonal = c(0, 1, 1))
+  expect_equal(validate_pmml(pmml(fit, exact_least_squares = TRUE), schema), 0)
+
+  fit <- Arima(JohnsonJohnson, order = c(2, 1, 3), seasonal = c(0, 1, 2))
+  expect_equal(validate_pmml(pmml(fit, exact_least_squares = TRUE), schema), 0)
+
+  fit <- Arima(AirPassengers, order = c(4, 2, 1), seasonal = c(1, 1, 1))
+  expect_equal(validate_pmml(pmml(fit, exact_least_squares = TRUE), schema), 0)
 })