+++ NEWS.md
@@ -1,9 +1,22 @@
+# pmml 2.2.0
+## Breaking Changes
+  * Deprecated parameters in the following functions have been removed, and replaced with equivalent in snake case.
+    - `pmml.iForest`: `anomalyThreshold` -> `anomaly_threshold`
+    - `pmml.kmeans`: `algorithm.name` -> `algorithm_name`
+    - `rename_wrap_var`: `wrap_data` -> `wrap_object`
+    - `xform_norm_discrete`: `inputVar` -> `input_var`
+
+## New Features
+* `pmml.ARIMA()` now exports prediction intervals for non-seasonal models.
+* `pmml.ARIMA()` can now export seasonal models with Exact Least Squares representation, and does so by default.
+
+
 # pmml 2.1.0
 
 ## Major Changes
 * Package now exports PMML with schema version 4.4.
 
-* `pmml.ARIMA()` function added - exports ARIMA time series models from the `forecast` package with conditional least squares forecasting.
+* `pmml.ARIMA()` function added - exports ARIMA time series models from the `forecast` package.
 
 
 ## Breaking Changes